Before:
```
$ bundle exec ruby -Iapp -Ilib -Itest test/unit/version_test.rb
...
Finished in 2.197927s, 36.8529 runs/s, 74.6158 assertions/s.

$ rake test
...
Finished in 73.659252s, 17.1194 runs/s, 32.2295 assertions/s
```

After:
```
$ bundle exec ruby -Iapp -Ilib -Itest test/unit/version_test.rb
...
Finished in 1.511227s, 53.5988 runs/s, 107.8594 assertions/s.

$ rake test 
...
Finished in 52.842494s, 23.8634 runs/s, 44.9070 assertions/s.
```

The gains in the suite are probably from fixing/removing the linkset creation. 

This test class still creates way too many records, but the next step will be refactoring test setup for the scope tests, which will be a lot more work than what I've done here!