@simi 2019-11-01
:thinking: I'm a little sceptical about this change. Faster test doesn't equal better test everytime.

:thinking: For example you call `@version.send(:update_prerelease)` in spec which is not testing actually this is called. There are also another potential callbacks which can differ this spec, but only this one will be called during this spec. IMHO the coverage scope of this spec is changed. Potential problems introduced by mistake in another on save callback will not be found in here.

:thinking: Also you had to remove `created_at` column assert as a trade-off (can't you ass it to `build ` as a argument?) for faster spec.

:thinking: The same works for changes in "sha" part where the test is actually called `be available from the database` and that was the reason for explicit `reload` usage in there. That spec doesn't make any sense now since database is not involved at all.