@nateberkopec 2019-11-01
Faster tests are more useful tests, because a slow test suite is difficult to run. If the tests do the same thing, we should prefer the faster way.

> For example you call @version.send(:update_prerelease) in spec which is not testing actually this is called.

It's a weird implementation. I think it could be changed. I understand why you're trying to create a prerelease *column*, but we should probably also override the accessor if prerelease is `nil`. Would you like me to do that as well?

> Also you had to remove created_at column assert as a trade-off (can't you ass it to build as a argument?) for faster spec.

I can add an `assert_nil` in this case or add a `build(created_at: arbitrary_datetime`, as you wish.

>  The same works for changes in "sha" part where the test is actually called be available from the database and that was the reason for explicit reload usage in there. That spec doesn't make any sense now since database is not involved at all.

Let's think more critically about what this is actually testing. The name doesn't make sense, true. I can change that if you wish. However, `sha256` is just a column. Should we save all columns individually to the database and then assert on their values to make sure that ActiveRecord works? I don't think that's useful, because Rails already tests that behavior extensively. Perhaps this test should be removed.

The other tests re: sha256_hex do not involve the database at all, because the implementations of those methods don't involve the DB.