@nateberkopec 2019-11-01
Pushed with those changes. I think switching create -> build revealed that `prerelease` depended on whether or not the record was saved to the database, which IMO is incorrect because we have the information required whether or not the record is saved.