@nateberkopec 2019-11-01
Ah, regarding `prerelease`, I'll have to think about that more. [ActiveRecord query methods](https://api.rubyonrails.org/classes/ActiveRecord/AttributeMethods/Query.html#method-i-query_attribute) depend on the attribute itself being set.