@simi 2019-11-03
> Faster tests are more useful tests, because a slow test suite is difficult to run.

:+1: Sure, there's nothing to discuss about this statement.

> If the tests do the same thing, we should prefer the faster way.

Actually this was the part I tried to point to. I wasn't sure all changes are keeping testing the same things.

For `sha256` tests, I can take a look. Maybe those tests are just some historic leftovers. Anyway if they just test column value now, there's no reason to keep them. Removed unneeded tests = fastest test!